Add support for HackerEarth and bug fixes
The plugin didn't work on SPOJ due to incorrect DOM access so fixed it.
The port.onMessage.addListener(function(arg){}) method of back.js required a JSON object as its argument
so changed prop.py to send messages to back.js in the form of json objects.
Added better exception handling with multiple try and except blocks and report exceptions to back.js